requirements.txt
2captcha-python==1.1.2
altgraph==0.17.2
anyio==3.6.1
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.9.3
blinker==1.5
Brotli==1.0.9
bs4==0.0.1
certifi==2022.6.15
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.0
cryptography==37.0.4
Faker==13.15.0
future==0.18.2
googlesearch-python==1.1.0
h11==0.13.0
h2==4.1.0
hpack==4.0.0
httpcore==0.15.0
httpx==0.23.0
hyperframe==6.0.1
idna==2.10
install==1.3.5
kaitaistruct==0.10
lxml==4.9.1
multipart==0.2.4
outcome==1.2.0
pefile==2022.5.30
pyasn1==0.4.8
pycparser==2.21
pydivert==2.1.0
pyinstaller==5.2
pyinstaller-hooks-contrib==2022.8
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyperclip==1.8.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pywin32-ctypes==0.2.0
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
selenium==4.3.0
selenium-wire==4.6.5
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spintax==1.0.4
trio==0.21.0
trio-websocket==0.9.2
uncurl==0.0.11
urllib3==1.26.10
webdriver-manager==3.8.1
wsproto==1.1.0
zstandard==0.18.0
